refactor: parseNumber를 parsers 폴더로 이동
